Move related files together to more easily understand what's going on
