+++ .eslintrc.json
@@ -11,12 +11,7 @@
   },
   "rules":{
     "no-console": "off",
-    "no-use-before-define": 0,
-    "react/jsx-filename-extension": [
-      1,
-      {
-        "extensions": [".js", ".jsx"]
-      }
-    ]
+    "no-use-before-define": ["error", { "functions": false }],
+    "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }]
   }
 }
