+++ src/components/Avatar/index.js
