+++ src/data/words.js
@@ -1,46 +1,26 @@
 export default [
-  'th',
   'and',
-  'to',
-  'of',
-  'a',
-  'i',
-  'in',
   'was',
-  'he',
   'that',
-  'it',
   'his',
   'her',
   'you',
-  'as',
   'had',
   'with',
   'for',
   'she',
   'not',
-  'at',
   'but',
-  'be',
-  'my',
-  'on',
   'have',
   'him',
-  'is',
   'said',
-  'me',
   'which',
-  'by',
-  'so',
   'this',
   'all',
   'from',
   'they',
-  'no',
   'were',
-  'if',
   'would',
-  'or',
   'when',
   'what',
   'there',
@@ -48,17 +28,12 @@ export default [
   'one',
   'could',
   'very',
-  'an',
   'who',
   'them',
-  'mr',
-  'we',
   'now',
   'more',
   'out',
-  'do',
   'are',
-  'up',
   'their',
   'your',
   'will',
@@ -87,7 +62,6 @@ export default [
   'other',
   'see',
   'must',
-  'am',
   'own',
   'come',
   'down',
@@ -102,7 +76,6 @@ export default [
   'great',
   'two',
   'can',
-  'go',
   'over',
   'too',
   'here',
@@ -119,7 +92,6 @@ export default [
   'though',
   'without',
   'went',
-  'us',
   'away',
   'day',
   'make',
@@ -201,7 +173,6 @@ export default [
   'perhaps',
   'sure',
   'heart',
-  'oh',
   'right',
   'against',
   'three',
@@ -321,7 +292,6 @@ export default [
   'person',
   'call',
   'given',
-  'i',
   'sometimes',
   'making',
   'short',
@@ -347,7 +317,6 @@ export default [
   'ought',
   'does',
   'known',
-  'it',
   'bed',
   'table',
   'that',
@@ -431,7 +400,6 @@ export default [
   'power',
   'seeing',
   'dead',
-  'i',
   'usual',
   'able',
   'second',
@@ -481,7 +449,6 @@ export default [
   'daughter',
   'stay',
   'forward',
-  'o',
   'interest',
   'thoughts',
   'followed',
@@ -543,7 +510,6 @@ export default [
   'meet',
   'instead',
   'wished',
-  'ah',
   'lived',
   'try',
   'red',
@@ -587,7 +553,6 @@ export default [
   'book',
   'appearance',
   'pay',
-  'i',
   'obliged',
   'particular',
   'pass',
@@ -633,7 +598,6 @@ export default [
   'stopped',
   'notice',
   'duty',
-  'he',
   'age',
   'figure',
   'leaving',
@@ -745,7 +709,6 @@ export default [
   'cry',
   'trust',
   'fast',
-  'ye',
   'warm',
   'danger',
   'trees',
@@ -815,7 +778,6 @@ export default [
   'mere',
   'farther',
   'changed',
-  'o',
   'passing',
   'girls',
   'force',
@@ -955,7 +917,6 @@ export default [
   'tired',
   'share',
   'pair',
-  'to',
   'aware',
   'colour',
   'writing',
@@ -972,7 +933,6 @@ export default [
   'worst',
   'legs',
   'watched',
-  'ay',
   'thee',
   'eight',
   'worthy',
@@ -987,7 +947,6 @@ export default [
   'degree',
   'shame',
   'places',
-  'ma',
   'couldn',
   'tongue',
   'according',
@@ -996,6 +955,5 @@ export default [
   'filled',
   'servants',
   'calling',
-  'fallen',
-  '368',
+  'fallen'
 ];
