+++ src/pages/Signup/index.js
@@ -4,7 +4,7 @@ import cx from 'classnames';
 import firebase from 'firebase/app';
 
 import { GoogleButton } from 'react-google-button';
-import { signinWithGoogle } from '../lib/google_signin';
+import { signinWithGoogle } from '../../lib/google_signin';
 
 class Signup extends React.Component {
   constructor(props) {
