+++ src/pages/TypingTest/__tests__/index.test.js
