+++ src/pages/TypingTest/components/Letter/__tests__/index.test.js
