+++ src/pages/TypingTest/components/Letter/index.js
@@ -4,10 +4,8 @@ import cx from 'classnames';
 
 import './style.css';
 
-const Text = ({ done, letter, isCurrent }) => (
-  <span
-    className={cx('letter', { done, current: isCurrent })}
-  >
+const Text = ({done, letter, isCurrent}) => (
+  <span className={cx('letter', {done, current: isCurrent})}>
     {letter}
   </span>
 );
@@ -15,7 +13,7 @@ const Text = ({ done, letter, isCurrent }) => (
 Text.propTypes = {
   done: PropTypes.bool.isRequired,
   letter: PropTypes.string.isRequired,
-  isCurrent: PropTypes.bool.isRequired,
+  isCurrent: PropTypes.bool.isRequired
 };
 
 export default Text;
