+++ src/pages/TypingTest/components/Text/__tests__/index.test.js
