+++ src/pages/TypingTest/components/Text/index.js
