+++ src/pages/TypingTest/components/Text/style.css
