+++ src/pages/TypingTest/components/TotalWords/index.js
