+++ src/pages/TypingTest/components/TotalWords/style.css
