+++ src/pages/TypingTest/style.css
